+++ art/src/main/java/top/skyeyefast/mchjong/art/SurvivalRecipes.java
@@ -18,6 +18,11 @@ final class SurvivalRecipes {
     private SurvivalRecipes() {}
 
     static void generate(GenerateData output) throws IOException {
+        output.write("data/mchjong/recipes/mahjong_manual.json", Map.of(
+            "type", "patchouli:shapeless_book_recipe", "book", "mchjong:guide",
+            "ingredients", List.of(item("book"), item("green_dye"), Map.of("item", "mchjong:mahjong_tile")),
+            "fabric:load_conditions", List.of(Map.of("condition", "fabric:all_mods_loaded", "values", List.of("patchouli"))),
+            "conditions", List.of(Map.of("type", "forge:mod_loaded", "modid", "patchouli"))));
         output.write("data/mchjong/recipes/dice.json", Map.of("type", "minecraft:crafting_shapeless",
             "category", "misc", "ingredients", List.of(item("bone_meal"), item("white_dye"), item("black_dye")),
             "result", stack("dice", 2, Map.of())));