+++ fabric/src/main/java/top/skyeyefast/mchjong/Mchjong.java
@@ -42,8 +42,11 @@ public class Mchjong implements ModInitializer {
             top.skyeyefast.mchjong.world.WorldSettings.of(server);
             top.skyeyefast.mchjong.config.ServerPresets.load(net.fabricmc.loader.api.FabricLoader.getInstance().getConfigDir());
         });
-        net.fabricmc.fabric.api.networking.v1.ServerPlayConnectionEvents.JOIN.register((handler, sender, server) ->
-            top.skyeyefast.mchjong.config.ServerPresets.send(handler.player));
+        net.fabricmc.fabric.api.networking.v1.ServerPlayConnectionEvents.JOIN.register((handler, sender, server) -> {
+            top.skyeyefast.mchjong.config.ServerPresets.send(handler.player);
+            if (net.fabricmc.loader.api.FabricLoader.getInstance().isModLoaded("patchouli"))
+                top.skyeyefast.mchjong.compat.patchouli.ManualGift.give(handler.player);
+        });
         top.skyeyefast.mchjong.recipe.MahjongRecipes.SERIALIZERS.forEach((name, serializer) ->
             Registry.register(BuiltInRegistries.RECIPE_SERIALIZER, MahjongContent.id(name), serializer));
         top.skyeyefast.mchjong.world.MahjongSounds.EVENTS.forEach((name, event) ->